fuse shared expert proj-mm
