+++ models/demos/deepseek_v3_b1/fused_ops/moe/moe_kernel.cpp
@@ -1,0 +1,625 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
+// SPDX-License-Identifier: Apache-2.0
+
+// Fused MoE kernel: Routed Expert + Shared Expert
+// Single kernel file, compiles correctly for all RISC cores.
+// Three-level structure mirrors Python-side: MoeOp > MoeRoutedExpertOp / MoeSharedExpertOp
+//   Core   → core role flags      (Core::Routed::*, Core::Shared::*)
+//   Moe    → compile-time args    (Moe::Routed::*, Moe::Shared::*)
+//
+// Routed expert pipeline:
+//   1. Mcast Input → 2. Gate Matmul+Sigmoid → 3. Gate Gather → 4. Gate TopK
+//   5. Index/Scale Mcast → 6. gate_proj (DRAM MM+SiLU) → 7. up_proj (DRAM MM)
+//   8. Fused mul → 9. down_proj Gather → 10. down_proj Mcast → 11. down_proj → 12. Add
+//
+// Shared expert pipeline (fused):
+//   3a. Gate/Up KN-sliced matmul (128 cores, overlaps with step 4)
+//
+// Optimizations:
+//   - gate_proj and up_proj share same CB (ResetCBIn1 between uses)
+//   - Input mcast shared between routed and shared expert
+
+#include "../../unified_kernels/kernel_op_api.hpp"
+#include "../../unified_kernels/kernel_utils.hpp"
+#include "../../unified_kernels/mcast.hpp"
+#include "../../unified_kernels/matmul.hpp"
+#include "../../unified_kernels/gather.hpp"
+#include "../../unified_kernels/deepseek_moe_gate.hpp"
+#include "../../unified_kernels/dram_streaming_matmul.hpp"
+#include "../../unified_kernels/eltwise_mul.hpp"
+#include "../../unified_kernels/eltwise_add.hpp"
+#include "../../unified_kernels/kn_sliced_matmul.hpp"
+
+// Compile-time role flags for dead code elimination via if constexpr.
+// Mirrors Python-side MoeRoutedExpertOp / MoeSharedExpertOp split.
+struct Core {
+    struct Routed {
+        static constexpr bool is_sender_core = get_named_compile_time_arg_val("is_sender_core") == 1;
+        static constexpr bool is_mcast_grid_core = get_named_compile_time_arg_val("is_mcast_grid_core") == 1;
+        static constexpr bool is_gate_mm_core = get_named_compile_time_arg_val("is_gate_mm_core") == 1;
+        static constexpr bool is_gate_proj_core = get_named_compile_time_arg_val("is_gate_proj_core") == 1;
+    };
+    struct Shared {
+        static constexpr bool is_compute_core = get_named_compile_time_arg_val("is_shared_compute_core") == 1;
+    };
+    // Combined: cores that receive the input mcast
+    static constexpr bool is_input_mcast_receiver =
+        Routed::is_gate_mm_core || Routed::is_gate_proj_core || Shared::is_compute_core;
+};
+
+void kernel_main() {
+// ============================================================================
+// Compile-time args — grouped into Moe::Routed / Moe::Shared structs
+// ============================================================================
+#if defined(COMPILE_FOR_NCRISC)
+
+    struct Moe {
+        struct Routed {
+            // Mcast (receiver)
+            using McastCTArgs = deepseek_b1_ops::Mcast::ReceiverCTArgs;
+            deepseek_b1_ops::Mcast::ReceiverArgs mcast_args{
+                get_named_compile_time_arg_val("mcast_data_receiver_semaphore"),
+                get_named_compile_time_arg_val("mcast_dst_cb"),
+                get_named_compile_time_arg_val("mcast_dst_num_pages"),
+            };
+
+            // Gate Matmul (reader — no-op)
+            using GateMMCTArgs = deepseek_b1_ops::Matmul::ReaderCTArgs;
+            deepseek_b1_ops::Matmul::ReaderArgs gate_mm_args{};
+
+            // Gather (sender)
+            deepseek_b1_ops::Gather::SenderArgs gather_args{
+                get_named_compile_time_arg_val("gather_dest_noc_x"),
+                get_named_compile_time_arg_val("gather_dest_noc_y"),
+                get_named_compile_time_arg_val("gather_data_size_bytes"),
+                get_named_compile_time_arg_val("gather_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("gather_src_cb"),
+                get_named_compile_time_arg_val("gather_src_num_pages"),
+                get_named_compile_time_arg_val("gather_sender_grid_start_x"),
+                get_named_compile_time_arg_val("gather_sender_grid_start_y"),
+                get_named_compile_time_arg_val("gather_sender_grid_end_x"),
+                get_named_compile_time_arg_val("gather_sender_grid_end_y"),
+                get_named_compile_time_arg_val("gather_row_major"),
+                get_named_compile_time_arg_val("gather_receiver_data_addr"),
+                0,  // sender_idx (unused when UsePerCoreSenderIdx=false)
+            };
+
+            // Gate (reader — no-op)
+            using GateCTArgs = deepseek_b1_ops::DeepseekMoeGate::ReaderCTArgs;
+
+            // Index Mcast (receiver)
+            deepseek_b1_ops::Mcast::ReceiverArgs index_mcast_args{
+                get_named_compile_time_arg_val("index_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("gate_proj_cb_index"),
+                get_named_compile_time_arg_val("index_mcast_num_pages"),
+            };
+
+            // Expert Scale Mcast (receiver)
+            deepseek_b1_ops::Mcast::ReceiverArgs expert_scale_mcast_args{
+                get_named_compile_time_arg_val("expert_scale_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("mul_cb_scalar_src"),
+                get_named_compile_time_arg_val("expert_scale_mcast_num_pages"),
+            };
+
+            // gate_proj DRAM Streaming Matmul (reader)
+            using GateProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ReaderCTArgs<
+                get_named_compile_time_arg_val("gate_proj_cb_in1"),
+                get_named_compile_time_arg_val("gate_proj_cb_out"),
+                get_named_compile_time_arg_val("gate_proj_in1_tensor_addr"),
+                get_named_compile_time_arg_val("gate_proj_in1_page_size"),
+                get_named_compile_time_arg_val("gate_proj_in1_num_pages"),
+                get_named_compile_time_arg_val("gate_proj_subblock_k"),
+                get_named_compile_time_arg_val("gate_proj_per_core_n"),
+                get_named_compile_time_arg_val("gate_proj_in1_block_size_bytes"),
+                get_named_compile_time_arg_val("gate_proj_out_num_tiles"),
+                get_named_compile_time_arg_val("gate_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("gate_proj_bank_id"),
+                get_named_compile_time_arg_val("gate_proj_vc"),
+                1,  // enable_indexing
+                get_named_compile_time_arg_val("gate_proj_cb_index"),
+                get_named_compile_time_arg_val("gate_proj_index_offset"),
+                get_named_compile_time_arg_val("use_hardcoded_expert_index")>;
+
+            // up_proj DRAM Streaming Matmul (reader) — shares CB with gate_proj
+            using UpProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ReaderCTArgs<
+                get_named_compile_time_arg_val("up_proj_cb_in1"),
+                get_named_compile_time_arg_val("up_proj_cb_mm_out"),
+                get_named_compile_time_arg_val("up_proj_in1_tensor_addr"),
+                get_named_compile_time_arg_val("up_proj_in1_page_size"),
+                get_named_compile_time_arg_val("up_proj_in1_num_pages"),
+                get_named_compile_time_arg_val("up_proj_subblock_k"),
+                get_named_compile_time_arg_val("up_proj_per_core_n"),
+                get_named_compile_time_arg_val("up_proj_in1_block_size_bytes"),
+                get_named_compile_time_arg_val("up_proj_out_num_tiles"),
+                get_named_compile_time_arg_val("up_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("up_proj_bank_id"),
+                get_named_compile_time_arg_val("up_proj_vc"),
+                1,  // enable_indexing
+                get_named_compile_time_arg_val("up_proj_cb_index"),
+                get_named_compile_time_arg_val("up_proj_index_offset"),
+                get_named_compile_time_arg_val("use_hardcoded_expert_index")>;
+
+            // Eltwise Mul (reader — no-op)
+            using MulCTArgs = deepseek_b1_ops::EltwiseMul::ReaderCTArgs;
+
+            // down_proj Gather (sender)
+            deepseek_b1_ops::Gather::SenderArgs down_proj_gather_args{
+                get_named_compile_time_arg_val("down_proj_gather_dest_noc_x"),
+                get_named_compile_time_arg_val("down_proj_gather_dest_noc_y"),
+                get_named_compile_time_arg_val("down_proj_gather_data_size_bytes"),
+                get_named_compile_time_arg_val("down_proj_gather_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("down_proj_gather_src_cb"),
+                get_named_compile_time_arg_val("down_proj_gather_src_num_pages"),
+                get_named_compile_time_arg_val("down_proj_gather_sender_grid_start_x"),
+                get_named_compile_time_arg_val("down_proj_gather_sender_grid_start_y"),
+                get_named_compile_time_arg_val("down_proj_gather_sender_grid_end_x"),
+                get_named_compile_time_arg_val("down_proj_gather_sender_grid_end_y"),
+                get_named_compile_time_arg_val("down_proj_gather_row_major"),
+                get_named_compile_time_arg_val("down_proj_gather_receiver_data_addr"),
+                get_named_compile_time_arg_val("down_proj_gather_sender_idx"),
+            };
+
+            // down_proj Mcast (receiver)
+            deepseek_b1_ops::Mcast::ReceiverArgs down_proj_mcast_args{
+                get_named_compile_time_arg_val("down_proj_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("down_proj_mcast_dst_cb"),
+                get_named_compile_time_arg_val("down_proj_mcast_dst_num_pages"),
+            };
+
+            // down_proj DRAM Streaming Matmul (reader)
+            using DownProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ReaderCTArgs<
+                get_named_compile_time_arg_val("down_proj_cb_in1"),
+                get_named_compile_time_arg_val("down_proj_cb_out"),
+                get_named_compile_time_arg_val("down_proj_in1_tensor_addr"),
+                get_named_compile_time_arg_val("down_proj_in1_page_size"),
+                get_named_compile_time_arg_val("down_proj_in1_num_pages"),
+                get_named_compile_time_arg_val("down_proj_subblock_k"),
+                get_named_compile_time_arg_val("down_proj_per_core_n"),
+                get_named_compile_time_arg_val("down_proj_in1_block_size_bytes"),
+                get_named_compile_time_arg_val("down_proj_out_num_tiles"),
+                get_named_compile_time_arg_val("down_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("down_proj_bank_id"),
+                get_named_compile_time_arg_val("down_proj_vc"),
+                1,  // enable_indexing
+                get_named_compile_time_arg_val("down_proj_cb_index"),
+                get_named_compile_time_arg_val("down_proj_index_offset"),
+                get_named_compile_time_arg_val("use_hardcoded_expert_index")>;
+
+            // Eltwise Add (reader — no-op)
+            using AddCTArgs = deepseek_b1_ops::EltwiseAdd::ReaderCTArgs;
+        } routed;
+
+        struct Shared {
+            // KN-sliced matmul (reader — no-op for NCRISC)
+            using GUMatmulCTArgs = deepseek_b1_ops::KNSlicedMatmul::ReaderCTArgs;
+            deepseek_b1_ops::KNSlicedMatmul::ReaderArgs gu_matmul_args{};
+        } shared;
+    } moe;
+
+    // Setup sharded persistent buffers (imperative — outside struct)
+    if constexpr (Core::Routed::is_sender_core) {
+        constexpr uint32_t mcast_src_cb = get_named_compile_time_arg_val("mcast_src_cb");
+        constexpr uint32_t mcast_src_num_pages = get_named_compile_time_arg_val("mcast_src_num_pages");
+        unified_kernels::setup_sharded_buffer(mcast_src_cb, mcast_src_num_pages);
+
+        constexpr uint32_t gate_bias_cb = get_named_compile_time_arg_val("gate_bias_cb");
+        constexpr uint32_t gate_input_indices_cb = get_named_compile_time_arg_val("gate_input_indices_cb");
+        unified_kernels::setup_sharded_buffer(gate_bias_cb, 1);
+        unified_kernels::setup_sharded_buffer(gate_input_indices_cb, 1);
+    }
+    if constexpr (Core::Routed::is_gate_mm_core) {
+        constexpr uint32_t gate_mm_in1 = get_named_compile_time_arg_val("gate_mm_in1");
+        constexpr uint32_t gate_mm_k_num_tiles = get_named_compile_time_arg_val("gate_mm_k_num_tiles");
+        constexpr uint32_t gate_mm_out_w = get_named_compile_time_arg_val("gate_mm_out_w");
+        unified_kernels::setup_sharded_buffer(gate_mm_in1, gate_mm_k_num_tiles * gate_mm_out_w);
+    }
+    if constexpr (Core::Routed::is_gate_proj_core) {
+        constexpr uint32_t mul_cb_in1 = get_named_compile_time_arg_val("mul_cb_in1");
+        constexpr uint32_t mul_num_tiles = get_named_compile_time_arg_val("mul_num_tiles");
+        unified_kernels::setup_sharded_buffer(mul_cb_in1, mul_num_tiles);
+
+        constexpr uint32_t add_cb_in0 = get_named_compile_time_arg_val("add_cb_in0");
+        constexpr uint32_t add_cb_in0_wait_tiles = get_named_compile_time_arg_val("add_cb_in0_wait_tiles");
+        unified_kernels::setup_sharded_buffer(add_cb_in0, add_cb_in0_wait_tiles);
+
+        constexpr uint32_t add_cb_in1 = get_named_compile_time_arg_val("add_cb_in1");
+        constexpr uint32_t add_cb_in1_wait_tiles = get_named_compile_time_arg_val("add_cb_in1_wait_tiles");
+        unified_kernels::setup_sharded_buffer(add_cb_in1, add_cb_in1_wait_tiles);
+    }
+    if constexpr (Core::Shared::is_compute_core) {
+        constexpr uint32_t shared_gu_weights_cb = get_named_compile_time_arg_val("shared_gu_weights_cb");
+        constexpr uint32_t shared_gu_weights_num_pages = get_named_compile_time_arg_val("shared_gu_weights_num_pages");
+        unified_kernels::setup_sharded_buffer(shared_gu_weights_cb, shared_gu_weights_num_pages);
+    }
+
+#elif defined(COMPILE_FOR_BRISC)
+
+    struct Moe {
+        struct Routed {
+            // Mcast (sender)
+            using McastCTArgs = deepseek_b1_ops::Mcast::SenderCTArgs<
+                get_named_compile_time_arg_val("mcast_num_cores"),
+                get_named_compile_time_arg_val("mcast_is_part_of_receiver_grid"),
+                Core::Routed::is_sender_core && Core::Routed::is_mcast_grid_core>;
+            deepseek_b1_ops::Mcast::SenderArgs mcast_args{
+                get_named_compile_time_arg_val("mcast_dest_noc_start_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_start_y"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_y"),
+                get_named_compile_time_arg_val("mcast_data_sender_semaphore"),
+                get_named_compile_time_arg_val("mcast_data_receiver_semaphore"),
+                get_named_compile_time_arg_val("mcast_data_size_bytes"),
+                get_named_compile_time_arg_val("mcast_src_cb"),
+                get_named_compile_time_arg_val("mcast_src_num_pages"),
+                get_read_ptr(get_named_compile_time_arg_val("mcast_src_cb")),
+                get_write_ptr(get_named_compile_time_arg_val("mcast_dst_cb")),
+            };
+
+            // Gate Matmul (writer — no-op)
+            using GateMMCTArgs = deepseek_b1_ops::Matmul::WriterCTArgs;
+            deepseek_b1_ops::Matmul::WriterArgs gate_mm_args{};
+
+            // Gather (receiver)
+            deepseek_b1_ops::Gather::ReceiverArgs gather_args{
+                get_named_compile_time_arg_val("gather_noc0_num_senders"),
+                get_named_compile_time_arg_val("gather_noc1_num_senders"),
+                get_named_compile_time_arg_val("gather_noc0_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("gather_noc1_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("gather_dst_cb"),
+                get_named_compile_time_arg_val("gather_dst_num_pages"),
+            };
+
+            // Gate (writer)
+            using GateCTArgs = deepseek_b1_ops::DeepseekMoeGate::WriterCTArgs<
+                get_named_compile_time_arg_val("gate_output_cb"),
+                get_named_compile_time_arg_val("gate_output_indices_cb")>;
+
+            // Index Mcast (sender)
+            deepseek_b1_ops::Mcast::SenderArgs index_mcast_args{
+                get_named_compile_time_arg_val("mcast_dest_noc_start_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_start_y"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_y"),
+                get_named_compile_time_arg_val("index_mcast_sender_semaphore"),
+                get_named_compile_time_arg_val("index_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("index_mcast_data_size_bytes"),
+                get_named_compile_time_arg_val("gate_output_indices_cb"),
+                get_named_compile_time_arg_val("index_mcast_num_pages"),
+                get_read_ptr(get_named_compile_time_arg_val("gate_output_indices_cb")),
+                get_write_ptr(get_named_compile_time_arg_val("gate_proj_cb_index")),
+            };
+
+            // Expert Scale Mcast (sender)
+            deepseek_b1_ops::Mcast::SenderArgs expert_scale_mcast_args{
+                get_named_compile_time_arg_val("mcast_dest_noc_start_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_start_y"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_y"),
+                get_named_compile_time_arg_val("expert_scale_mcast_sender_semaphore"),
+                get_named_compile_time_arg_val("expert_scale_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("expert_scale_mcast_data_size_bytes"),
+                get_named_compile_time_arg_val("gate_output_cb"),
+                get_named_compile_time_arg_val("expert_scale_mcast_num_pages"),
+                get_read_ptr(get_named_compile_time_arg_val("gate_output_cb")),
+                get_write_ptr(get_named_compile_time_arg_val("mul_cb_scalar_src")),
+            };
+
+            // DRAM Streaming Matmul (writer — no-op for BRISC)
+            using GateProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::WriterCTArgs;
+            using UpProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::WriterCTArgs;
+
+            // Eltwise Mul (writer)
+            using MulCTArgs = deepseek_b1_ops::EltwiseMul::WriterCTArgs<
+                get_named_compile_time_arg_val("mul_cb_out"),
+                get_named_compile_time_arg_val("mul_num_tiles"),
+                get_named_compile_time_arg_val("mul_cb_scalar"),
+                get_named_compile_time_arg_val("mul_cb_scalar_src"),
+                get_named_compile_time_arg_val("mul_scalar_index_offset")>;
+
+            // down_proj Gather (receiver)
+            deepseek_b1_ops::Gather::ReceiverArgs down_proj_gather_args{
+                get_named_compile_time_arg_val("down_proj_gather_noc0_num_senders"),
+                get_named_compile_time_arg_val("down_proj_gather_noc1_num_senders"),
+                get_named_compile_time_arg_val("down_proj_gather_noc0_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("down_proj_gather_noc1_receiver_semaphore_id"),
+                get_named_compile_time_arg_val("down_proj_gather_dst_cb"),
+                get_named_compile_time_arg_val("down_proj_gather_dst_num_pages"),
+            };
+
+            // down_proj Mcast (sender)
+            deepseek_b1_ops::Mcast::SenderArgs down_proj_mcast_args{
+                get_named_compile_time_arg_val("mcast_dest_noc_start_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_start_y"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_x"),
+                get_named_compile_time_arg_val("mcast_dest_noc_end_y"),
+                get_named_compile_time_arg_val("down_proj_mcast_sender_semaphore"),
+                get_named_compile_time_arg_val("down_proj_mcast_receiver_semaphore"),
+                get_named_compile_time_arg_val("down_proj_mcast_data_size_bytes"),
+                get_named_compile_time_arg_val("down_proj_mcast_src_cb"),
+                get_named_compile_time_arg_val("down_proj_mcast_src_num_pages"),
+                get_read_ptr(get_named_compile_time_arg_val("down_proj_mcast_src_cb")),
+                get_write_ptr(get_named_compile_time_arg_val("down_proj_mcast_dst_cb")),
+            };
+
+            // down_proj DRAM Matmul (writer — no-op for BRISC)
+            using DownProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::WriterCTArgs;
+
+            // Eltwise Add (writer — no-op)
+            using AddCTArgs = deepseek_b1_ops::EltwiseAdd::WriterCTArgs;
+        } routed;
+
+        struct Shared {
+            // KN-sliced matmul (writer — no-op for BRISC)
+            using GUMatmulCTArgs = deepseek_b1_ops::KNSlicedMatmul::WriterCTArgs;
+            deepseek_b1_ops::KNSlicedMatmul::WriterArgs gu_matmul_args{};
+        } shared;
+    } moe;
+
+#elif defined(COMPILE_FOR_TRISC)
+
+    struct Moe {
+        struct Routed {
+            // Mcast (compute — no-op)
+            using McastCTArgs = deepseek_b1_ops::Mcast::ComputeCTArgs;
+            deepseek_b1_ops::Mcast::ComputeArgs mcast_args{};
+
+            // Gate Matmul (compute)
+            using GateMMCTArgs = deepseek_b1_ops::Matmul::ComputeCTArgs<
+                get_named_compile_time_arg_val("gate_mm_out_w"),
+                false,  // transpose
+                get_named_compile_time_arg_val("gate_mm_fused_activation")>;
+            deepseek_b1_ops::Matmul::ComputeArgs gate_mm_args{
+                get_named_compile_time_arg_val("gate_mm_in0"),
+                get_named_compile_time_arg_val("gate_mm_in1"),
+                get_named_compile_time_arg_val("gate_mm_out"),
+                get_named_compile_time_arg_val("gate_mm_k_num_tiles"),
+            };
+
+            // Gather (compute — no-op)
+            deepseek_b1_ops::Gather::ComputeArgs gather_args{};
+
+            // Gate (compute)
+            using GateCTArgs = deepseek_b1_ops::DeepseekMoeGate::ComputeCTArgs<
+                get_named_compile_time_arg_val("gate_input_cb"),
+                get_named_compile_time_arg_val("gate_bias_cb"),
+                get_named_compile_time_arg_val("gate_input_indices_cb"),
+                get_named_compile_time_arg_val("gate_output_cb"),
+                get_named_compile_time_arg_val("gate_output_indices_cb"),
+                get_named_compile_time_arg_val("gate_eps"),
+                get_named_compile_time_arg_val("gate_scaling_factor"),
+                get_named_compile_time_arg_val("gate_enable_sigmoid")>;
+
+            // Index Mcast (compute — no-op)
+            deepseek_b1_ops::Mcast::ComputeArgs index_mcast_args{};
+
+            // Expert Scale Mcast (compute — no-op)
+            deepseek_b1_ops::Mcast::ComputeArgs expert_scale_mcast_args{};
+
+            // gate_proj DRAM Streaming Matmul (compute)
+            using GateProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ComputeCTArgs<
+                get_named_compile_time_arg_val("gate_proj_cb_in0"),
+                get_named_compile_time_arg_val("gate_proj_cb_in1"),
+                get_named_compile_time_arg_val("gate_proj_cb_out"),
+                get_named_compile_time_arg_val("gate_proj_subblock_k"),
+                get_named_compile_time_arg_val("gate_proj_per_core_n"),
+                get_named_compile_time_arg_val("gate_proj_subblock_w"),
+                get_named_compile_time_arg_val("gate_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("gate_proj_tile_r_dim"),
+                get_named_compile_time_arg_val("gate_proj_fuse_silu"),
+                get_named_compile_time_arg_val("gate_proj_fp32_dest_acc_en")>;
+
+            // up_proj DRAM Streaming Matmul (compute) — shares CB with gate_proj
+            using UpProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ComputeCTArgs<
+                get_named_compile_time_arg_val("up_proj_cb_in0"),
+                get_named_compile_time_arg_val("up_proj_cb_in1"),
+                get_named_compile_time_arg_val("up_proj_cb_mm_out"),
+                get_named_compile_time_arg_val("up_proj_subblock_k"),
+                get_named_compile_time_arg_val("up_proj_per_core_n"),
+                get_named_compile_time_arg_val("up_proj_subblock_w"),
+                get_named_compile_time_arg_val("up_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("up_proj_tile_r_dim"),
+                get_named_compile_time_arg_val("up_proj_fuse_silu"),
+                get_named_compile_time_arg_val("up_proj_fp32_dest_acc_en")>;
+
+            // Eltwise Mul (compute)
+            using MulCTArgs = deepseek_b1_ops::EltwiseMul::ComputeCTArgs<
+                get_named_compile_time_arg_val("mul_cb_in0"),
+                get_named_compile_time_arg_val("mul_cb_in1"),
+                get_named_compile_time_arg_val("mul_cb_out"),
+                get_named_compile_time_arg_val("mul_num_tiles"),
+                get_named_compile_time_arg_val("up_proj_cb_mm_out"),   // cb_in0_wait
+                get_named_compile_time_arg_val("up_proj_per_core_n"),  // cb_in0_wait tiles
+                get_named_compile_time_arg_val("mul_cb_scalar"),
+                get_named_compile_time_arg_val("mul_fp32_dest_acc_en")>;
+
+            // down_proj Gather (compute — no-op)
+            deepseek_b1_ops::Gather::ComputeArgs down_proj_gather_args{};
+
+            // down_proj Mcast (compute — no-op)
+            deepseek_b1_ops::Mcast::ComputeArgs down_proj_mcast_args{};
+
+            // down_proj DRAM Streaming Matmul (compute)
+            using DownProjCTArgs = deepseek_b1_ops::DRAMStreamingMatmul::ComputeCTArgs<
+                get_named_compile_time_arg_val("down_proj_cb_in0"),
+                get_named_compile_time_arg_val("down_proj_cb_in1"),
+                get_named_compile_time_arg_val("down_proj_cb_out"),
+                get_named_compile_time_arg_val("down_proj_subblock_k"),
+                get_named_compile_time_arg_val("down_proj_per_core_n"),
+                get_named_compile_time_arg_val("down_proj_subblock_w"),
+                get_named_compile_time_arg_val("down_proj_num_subblocks_k"),
+                get_named_compile_time_arg_val("down_proj_tile_r_dim"),
+                get_named_compile_time_arg_val("down_proj_fuse_silu"),
+                get_named_compile_time_arg_val("down_proj_fp32_dest_acc_en")>;
+
+            // Eltwise Add (compute)
+            using AddCTArgs = deepseek_b1_ops::EltwiseAdd::ComputeCTArgs<
+                get_named_compile_time_arg_val("add_cb_in0"),
+                get_named_compile_time_arg_val("add_cb_in1"),
+                get_named_compile_time_arg_val("add_cb_out"),
+                get_named_compile_time_arg_val("add_num_tiles"),
+                get_named_compile_time_arg_val("add_cb_in0"),  // cb_in0_wait = cb_in0
+                get_named_compile_time_arg_val("add_cb_in0_wait_tiles"),
+                get_named_compile_time_arg_val("add_cb_in1_wait_tiles"),
+                get_named_compile_time_arg_val("add_sender_index"),
+                get_named_compile_time_arg_val("add_slice_size_bytes")>;
+        } routed;
+
+        struct Shared {
+            // KN-sliced matmul (compute)
+            using GUMatmulCTArgs = deepseek_b1_ops::KNSlicedMatmul::ComputeCTArgs<>;
+            deepseek_b1_ops::KNSlicedMatmul::ComputeArgs gu_matmul_args{
+                get_named_compile_time_arg_val("shared_gu_act_cb"),
+                get_named_compile_time_arg_val("shared_gu_weights_cb"),
+                get_named_compile_time_arg_val("shared_gu_out_cb"),
+                get_named_compile_time_arg_val("shared_gu_k_offset"),
+                get_named_compile_time_arg_val("shared_gu_k_per_core"),
+                get_named_compile_time_arg_val("shared_gu_act_total_tiles"),
+            };
+        } shared;
+    } moe;
+
+#endif
+
+    // ============================================================================
+    // Operation calls — Moe::Routed::* for types, moe.routed.* for args
+    // ============================================================================
+
+    // 1. Mcast Input: Broadcast input from sender core to all receiver cores
+    deepseek_b1_ops::Mcast::Op<
+        Moe::Routed::McastCTArgs,
+        Core::Routed::is_sender_core,
+        Core::Routed::is_mcast_grid_core,
+        Core::is_input_mcast_receiver,
+        true>
+        mcast;
+    mcast.init(moe.routed.mcast_args);
+    {
+        DeviceZoneScopedN("MCAST");
+        mcast(moe.routed.mcast_args);
+    }
+
+    // 2. Matmul + Activation: Routing matmul on gate_mm cores
+    {
+        DeviceZoneScopedN("MATMUL");
+        deepseek_b1_ops::Matmul::Op<Moe::Routed::GateMMCTArgs, Core::Routed::is_gate_mm_core, false, false> gate_mm;
+        gate_mm(moe.routed.gate_mm_args);
+    }
+
+    // 3. Gather: Collect matmul outputs from compute cores to sender core
+    {
+        DeviceZoneScopedN("GATHER");
+        deepseek_b1_ops::Gather::Op<Core::Routed::is_gate_mm_core, Core::Routed::is_sender_core, true> gather;
+        gather(moe.routed.gather_args);
+    }
+
+    // 3a. Shared Expert: Gate/Up KN-sliced matmul on 128 compute cores
+    //     Overlaps with Gate Compute (#4) — 128 cores do matmul while sender does TopK
+    //     pop_act=false (gate_proj cores still need input for DRAM streaming matmul)
+    {
+        DeviceZoneScopedN("SHARED_GATE_UP_MATMUL");
+        deepseek_b1_ops::KNSlicedMatmul::Op<Moe::Shared::GUMatmulCTArgs, Core::Shared::is_compute_core, false, false>
+            shared_gu_matmul;
+        shared_gu_matmul(moe.shared.gu_matmul_args);
+    }
+
+    // 4. Gate: Top-K expert selection (on sender core only)
+    {
+        DeviceZoneScopedN("GATE");
+        deepseek_b1_ops::DeepseekMoeGate::Op<Moe::Routed::GateCTArgs, Core::Routed::is_sender_core> gate;
+        gate();
+    }
+
+    // 5. Mcast Index: Broadcast expert indices to compute cores
+    {
+        DeviceZoneScopedN("MCAST_INDEX");
+        mcast(moe.routed.index_mcast_args);
+    }
+
+    // 5b. Mcast Expert Scale: Broadcast expert scale to gate_proj cores
+    {
+        DeviceZoneScopedN("MCAST_EXPERT_SCALE");
+        deepseek_b1_ops::Mcast::Op<
+            Moe::Routed::McastCTArgs,
+            Core::Routed::is_sender_core,
+            Core::Routed::is_mcast_grid_core,
+            Core::Routed::is_gate_proj_core,
+            true,  // pop_src
+            true>  // UpdateSemaphoreAddr
+            expert_scale_mcast;
+        expert_scale_mcast(moe.routed.expert_scale_mcast_args);
+    }
+
+    // 6. gate_proj: DRAM Streaming Matmul + SiLU (PopIn0=false, keep input for up_proj)
+    {
+        DeviceZoneScopedN("GATE_PROJ");
+        constexpr uint32_t gate_proj_cb_in1_addr = get_named_compile_time_arg_val("gate_proj_in1_buf_addr");
+        deepseek_b1_ops::DRAMStreamingMatmul::
+            Op<Moe::Routed::GateProjCTArgs, Core::Routed::is_gate_proj_core, false, true, gate_proj_cb_in1_addr>
+                gate_proj_mm;
+        gate_proj_mm();
+    }
+
+    // 7. up_proj: DRAM Streaming Matmul (PopIn0=true, ResetCBIn1=true — shared CB with gate_proj)
+    {
+        DeviceZoneScopedN("UP_PROJ");
+        constexpr uint32_t cb_in1_addr = get_named_compile_time_arg_val("gate_proj_in1_buf_addr");
+        deepseek_b1_ops::DRAMStreamingMatmul::
+            Op<Moe::Routed::UpProjCTArgs, Core::Routed::is_gate_proj_core, true, true, cb_in1_addr>
+                up_proj;
+        up_proj();
+    }
+
+    // 8. Mul: Element-wise multiply (up_proj * gate_proj * expert_scale)
+    {
+        DeviceZoneScopedN("MUL");
+        deepseek_b1_ops::EltwiseMul::Op<Moe::Routed::MulCTArgs, Core::Routed::is_gate_proj_core> mul_op;
+        mul_op();
+    }
+
+    // 9. down_proj Gather: Gather fused output from gate_proj cores to sender core
+    {
+        DeviceZoneScopedN("DOWN_PROJ_GATHER");
+        deepseek_b1_ops::Gather::Op<Core::Routed::is_gate_proj_core, Core::Routed::is_sender_core, true, true>
+            down_proj_gather;
+        down_proj_gather(moe.routed.down_proj_gather_args);
+    }
+
+    // 10. down_proj Mcast: Broadcast gathered fused output to gate_proj cores
+    {
+        DeviceZoneScopedN("DOWN_PROJ_MCAST");
+        deepseek_b1_ops::Mcast::Op<
+            Moe::Routed::McastCTArgs,
+            Core::Routed::is_sender_core,
+            Core::Routed::is_mcast_grid_core,
+            Core::Routed::is_gate_proj_core,
+            true,  // pop_src
+            true>  // UpdateSemaphoreAddr
+            down_proj_mcast;
+        down_proj_mcast(moe.routed.down_proj_mcast_args);
+    }
+
+    // 11. down_proj: DRAM Streaming Matmul
+    {
+        DeviceZoneScopedN("DOWN_PROJ");
+        constexpr uint32_t down_proj_cb_in1_addr = get_named_compile_time_arg_val("down_proj_in1_buf_addr");
+        deepseek_b1_ops::DRAMStreamingMatmul::
+            Op<Moe::Routed::DownProjCTArgs, Core::Routed::is_gate_proj_core, true, true, down_proj_cb_in1_addr>
+                down_proj;
+        down_proj();
+    }
+
+    // 12. Eltwise Add: down_proj + fused_add
+    {
+        DeviceZoneScopedN("ELTWISE_ADD");
+        deepseek_b1_ops::EltwiseAdd::Op<Moe::Routed::AddCTArgs, Core::Routed::is_gate_proj_core> add_op;
+        add_op();
+    }
+
+    // Teardown (one teardown since all mcasts reuse the same semaphores)
+    mcast.teardown();
+
+#if defined(COMPILE_FOR_NCRISC) || defined(COMPILE_FOR_BRISC)
+    noc_async_write_barrier();
+    noc_async_atomic_barrier();
+#endif
+}
